chore: house cleaning

+++ worker.ts
@@ -1,4 +1,4 @@
-import { NPage, Website, pageTitle } from "@spider-rs/spider-rs";
+import { type NPage, Website, pageTitle } from "@spider-rs/spider-rs";
 import { db } from "./database";
 import Crawler from "crawler";
 const process = require("node:process");
@@ -24,12 +24,7 @@ async function run() {
 
 	// optional: page event handler
 	const onPageEvent = async (_err, page: NPage) => {
-		// const p = await page.fetch();
-		// const html = p.getHtml();
-		// console.log(html);
-		// console.log(convert(page.content))
 		const title = pageTitle(page);
-		console.info(`Title of ${page.url} is '${title}'`);
 		if (page.statusCode === 200) {
 			crawledPages.push({
 				content: convert(page.content),
@@ -38,58 +33,17 @@ async function run() {
 				domain: domain,
 			})
 		}
-
-		// await DB.any(
-		// 	`
-		// 	INSERT INTO websites (title, url, domain, content) VALUES($1:raw, $2, $3, $4:raw)
-		// 	ON CONFLICT (url) DO UPDATE SET content = EXCLUDED.content
-		// 	`,
-		// 	["$$" + title + "$$", page.url, domain, "$$" + convert(page.content) + "$$"],
-		// );
-
-		// website.pushData({
-		// 	status: page.statusCode,
-		// 	html: convert(page.content),
-		// 	url: page.url,
-		// 	title,
-		// 	domain: domain,
-		// });
 	};
 
 	await website.crawl(onPageEvent, false, true);
 
-	await db.collection("websites").deleteMany({ domain: domain })
-	await db.collection("websites").insertMany(crawledPages)
-	// console.log(website.readData())
-	// await website.exportJsonlData(`./storage/${domain}.jsonl`);
-	// console.log(website.getLinks());
+	if (crawledPages.length > 0) {
+		await db.collection("websites").deleteMany({ domain: domain })
+		await db.collection("websites").insertMany(crawledPages)
+	}
 
-	process.exit(0);
-	// process.send({ type: "done", domain: domain });
+	// process.exit(0);
+	process.send({ type: "done", domain: domain });
 }
 
 run();
-
-async function run2() {
-	const c = new Crawler({
-		maxConnections: 20,
-		rateLimit: 1000,
-		// This will be called for each crawled page
-		callback: (error, res, done) => {
-			if (error) {
-				console.log(error);
-			} else {
-				const $ = res.$;
-				// $ is Cheerio by default
-				//a lean implementation of core jQuery designed specifically for the server
-				console.log($("title").text());
-				console.log($("body").text());
-			}
-			console.log("Crawling done!");
-			// done();
-		},
-	});
-	c.add(domain);
-}
-
-// run2()
